feat: topic set/get methods
Implements base functionality for setting and getting key/values for a
topic.

When setting a key, if the topic does not exist, it will be created
first. Creating a topic splits the key by separator and defines columns
based on each fragment. The names are anonymous because we cannot infer
a generic name based on the key.

Once the topic is created, setting a key following the same method as
creating topic columns. the columns are populated with the key
fragments, and the value column with the received parameter value.

Getting a value in a topic by key, splits the key up, and uses the key
fragments as where clauses. The intact key and value are returned
as entries `[[key, value], ...]`.

Getting a key can be done with wildcards `*` in place of any key
fragment. Multiple fragments can be wildcards as well to create
multidimensional queries. Providing a wildcard as _the_ key, returns
available entries for a topic.