refactor!: remove singleton functionality
To expose control to change the file of the connection, export a factory
function to create the instance of the store. This takes a config and
abstracts away the initialization of the store itself.
This allows us to have multiple separate stores, including mixing file
and in-memory versions, each with separate topics.