refactor(proj): replace standard with eslint
I'm worried about the StandardJS project at the moment. There are some
movements to get maintenance and governence in place.

This was also a sticky choice because Eslint v9 just released, however
there are compatibility issues with plugins like `import`. Hopefully in
the future we can use v9 but for now, sticking with v8. Another bonus to
v8 is that we can still have our config in package.json.

The rules defined were taken from the `eslint-config-standard` package,
then ported over to stylistic using the available migration script.