feat: remove @databases dependency
The internals of the application use @databases/sqlite-sync, which has a
dependency on `better-sqlite3`. The convienience of @databases is great
for larger projects, applications, and is a fantastic abstraction.

We can can reduce our dependencies by removing @databases and using
`better-sqlite3` directly. It's a little more verbose, but manageable
for inside this library.